scroll bug // entryList => contentList flash of previous items
In the contentList, if you scroll further than the maximum height of the list. Then move back to the entryList then back into the contentList. The list would still be scrolled down to where it was the last time when it was in the contentList.

Flash was because the list got updated async while the list got displayed instantly. simply added an await to fix this